Changes: - Added several corner cases service in open api - Moved all endpoints to one controller - Added integer cast exception handler - Removed AnimalsController.java - Introduced version 2.0.0 of API
